Starting to convert the old conversation view controller to a conversation THREAD view controller.
